修复dp计算导致崩溃的问题

+++ app/src/main/java/com/qihuan/photowidget/bean/WidgetInfo.kt
@@ -14,7 +14,7 @@ data class WidgetInfo(
     @PrimaryKey
     val widgetId: Int,
     var uri: Uri,
-    val verticalPadding: Int,
-    val horizontalPadding: Int,
-    val widgetRadius: Int,
+    val verticalPadding: Float,
+    val horizontalPadding: Float,
+    val widgetRadius: Float,
 )
